feat(llm): move the LLM chat special notes into core
The rest of the special-notes family already lives in core and is served
from the shared route table; the five LLM chat routes were the only
holdouts, so the standalone worker answered none of them — the sidebar
chat could not create a chat, restore the last one, list history, or save
one out of the hidden subtree.

Nothing about them needed the server: the service used becca, date notes,
hoisted notes, search and i18next, all of which core has. Merging it into
core's special_notes also retires a verbatim copy of
getMonthlyParentNoteId, and puts getLlmChatHome next to the search and
SQL-console home lookups it was modelled on.

The specs merge rather than move, since core already had one for each
file. The route coverage now runs through CoreApiTester against the real
route table instead of calling the handlers directly, so registration and
the limit query-param parsing are covered too — and both suites run it,
against better-sqlite3 and against sql.js.
