+++ packages/trilium-core/src/routes/api/special_notes.ts
@@ -107,6 +107,29 @@ function createOrUpdateScriptLauncherFromApi(req: Request) {
     return specialNotesService.createOrUpdateScriptLauncherFromApi(req.body);
 }
 
+function createLlmChat() {
+    return specialNotesService.createLlmChat();
+}
+
+function getMostRecentLlmChat() {
+    const chat = specialNotesService.getMostRecentLlmChat();
+    // Return null explicitly if no chat found (not undefined)
+    return chat || null;
+}
+
+function getOrCreateLlmChat() {
+    return specialNotesService.getOrCreateLlmChat();
+}
+
+function getRecentLlmChats(req: Request) {
+    const limit = parseInt(req.query.limit as string) || 10;
+    return specialNotesService.getRecentLlmChats(limit);
+}
+
+function saveLlmChat(req: Request<{ llmChatNoteId: string }>) {
+    return specialNotesService.saveLlmChat(req.body?.llmChatNoteId);
+}
+
 export default {
     getInboxNote,
     getDayNote,
@@ -122,5 +145,10 @@ export default {
     saveSearchNote,
     createLauncher,
     resetLauncher,
-    createOrUpdateScriptLauncherFromApi
+    createOrUpdateScriptLauncherFromApi,
+    createLlmChat,
+    getMostRecentLlmChat,
+    getOrCreateLlmChat,
+    getRecentLlmChats,
+    saveLlmChat
 };
