+++ packages/trilium-core/src/services/special_notes.ts
@@ -10,7 +10,7 @@ import SearchContext from "./search/search_context.js";
 import { LBTPL_NOTE_LAUNCHER, LBTPL_CUSTOM_WIDGET, LBTPL_SPACER, LBTPL_SCRIPT } from "./hidden_subtree.js";
 import { t } from "i18next";
 import BNote from '../becca/entities/bnote.js';
-import { SaveSearchNoteResponse, SaveSqlConsoleResponse } from "@triliumnext/commons";
+import { SaveLlmChatResponse, SaveSearchNoteResponse, SaveSqlConsoleResponse } from "@triliumnext/commons";
 
 function getInboxNote(date: string) {
     const workspaceNote = hoistedNoteService.getWorkspaceNote();
@@ -276,6 +276,117 @@ function createOrUpdateScriptLauncherFromApi(opts: { id: string; title: string; 
     return launcherNote;
 }
 
+/**
+ * Gets the most recently modified LLM chat note.
+ * Used by sidebar chat to persist conversation across page refreshes.
+ * Returns null if no chat exists.
+ */
+function getMostRecentLlmChat() {
+    // Search for all llmChat notes and return the most recently modified
+    const results = searchService.searchNotes(
+        "note.type = llmChat",
+        new SearchContext({
+            ancestorNoteId: "_llmChat",
+            limit: 1,
+            orderBy: "utcDateModified",
+            orderDirection: "desc"
+        })
+    );
+
+    return results.length > 0 ? results[0] : null;
+}
+
+/**
+ * Gets the most recent LLM chat or creates a new one if none exists.
+ * Used by sidebar chat for persistent conversations.
+ */
+function getOrCreateLlmChat() {
+    const existingChat = getMostRecentLlmChat();
+
+    if (existingChat) {
+        return existingChat;
+    }
+
+    return createLlmChat();
+}
+
+/**
+ * Gets a list of recent LLM chat notes.
+ * Used by sidebar chat history popup.
+ */
+function getRecentLlmChats(limit: number = 10) {
+    const results = searchService.searchNotes(
+        "note.type = llmChat",
+        new SearchContext({
+            ancestorNoteId: "_llmChat",
+            limit,
+            orderBy: "utcDateModified",
+            orderDirection: "desc"
+        })
+    );
+
+    return results.map(note => ({
+        noteId: note.noteId,
+        title: note.title,
+        dateModified: note.utcDateModified
+    }));
+}
+
+function createLlmChat() {
+    const { note } = noteService.createNewNote({
+        parentNoteId: getMonthlyParentNoteId("_llmChat", "llmChat"),
+        title: `${t("special_notes.llm_chat_prefix")} ${dateUtils.localNowDateTime()}`,
+        content: JSON.stringify({
+            version: 1,
+            messages: []
+        }),
+        type: "llmChat",
+        mime: "application/json"
+    });
+
+    note.setLabel("iconClass", "bx bx-message-square-dots");
+    note.setLabel("keepCurrentHoisting");
+
+    return note;
+}
+
+function getLlmChatHome() {
+    const workspaceNote = hoistedNoteService.getWorkspaceNote();
+    if (!workspaceNote) {
+        throw new Error("Unable to find workspace note");
+    }
+
+    if (!workspaceNote.isRoot()) {
+        return workspaceNote.searchNoteInSubtree("#workspaceLlmChatHome") || workspaceNote.searchNoteInSubtree("#llmChatHome") || workspaceNote;
+    }
+    const today = dateUtils.localNowDate();
+
+    return workspaceNote.searchNoteInSubtree("#llmChatHome") || dateNoteService.getDayNote(today);
+}
+
+function saveLlmChat(llmChatNoteId: string | null) {
+    if (!llmChatNoteId) {
+        throw new Error(`Missing chat note ID`);
+    }
+
+    const llmChatNote = becca.getNote(llmChatNoteId);
+    if (!llmChatNote) {
+        throw new Error(`Unable to find LLM chat note ID: ${llmChatNoteId}`);
+    }
+
+    const llmChatHome = getLlmChatHome();
+
+    const result = llmChatNote.cloneTo(llmChatHome.noteId);
+
+    for (const parentBranch of llmChatNote.getParentBranches()) {
+        if (parentBranch.parentNote?.hasAncestor("_hidden")) {
+            parentBranch.markAsDeleted();
+        }
+    }
+
+    return result satisfies SaveLlmChatResponse;
+}
+
 export default {
     getInboxNote,
     createSqlConsole,
@@ -284,5 +395,10 @@ export default {
     saveSearchNote,
     createLauncher,
     resetLauncher,
-    createOrUpdateScriptLauncherFromApi
+    createOrUpdateScriptLauncherFromApi,
+    createLlmChat,
+    getMostRecentLlmChat,
+    getOrCreateLlmChat,
+    getRecentLlmChats,
+    saveLlmChat
 };
